All five endpoints done (missing pagination, fix exception)

+++ ecommerce/src/main/kotlin/it/polito/ecommerce/repositories/CustomerRepository.kt
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface CustomerRepository:CrudRepository<Customer, Int?> {
+interface CustomerRepository:CrudRepository<Customer, Long?> {
 
 }
